!29 #I6CYPE 实现MIN函数, 发现avx2的max比对的bug,NaN比任何float都大,也比任何float都小 * #I6CYPE 实现MIN函数 * 修复avx2 max的bug * vek的float数组最大值对比有bug, NaN比任何float都大, 也比任何float都小, 所以最大值比对的时候, 正确的应该是Na…

+++ generic_type.go
@@ -50,7 +50,7 @@ func findTypeByString(arr []string) (Type, error) {
 		}
 		if f, err := strconv.ParseFloat(str, 64); err == nil {
 			hasFloats = true
-			if f < MaxFloat32 {
+			if float32(f) < MaxFloat32 {
 				if mustFloat64(float32(f)) {
 					useFloat64 = true
 				} else {
